pusher/requirements/base.txt
aiokafka==0.12.0
websockets==15.0.1
pyjwt==2.10.1